Refactor utils in a folder
