+++ .eslintrc.json
@@ -12,6 +12,7 @@
   },
   "rules": {
     "import/no-unresolved": [2, { "commonjs": true, "amd": true }],
+    "import/named": [2, { "commonjs": true }],
     "indent": ["error", 2, { "SwitchCase": 1 }],
     "linebreak-style": ["error", "unix"],
     "quotes": ["error", "double"],
