+++ src/mergeTiles/getCoordBoundary.js
