+++ src/mergeTiles/getInnerPoints.js
